Change width to height for horizontal seam removal.
Previously images with height > width were getting chopped off
after the first seam carving. The image was being cropped height-
wise to its width and the rest of the image was written out as
black. This caused the energy function to just remove the bottom
black pixels on n=2 and subsequent runs with horizontal seams.